image_cuda_compatible and image classes use float array to store the image, instead of unsigned short
Raw images are unsigned shorts, but i realized it is more handy to handle the mas float, since
that way i can sum two images etc...

+++ cuda_image_kernel_calls.cu
@@ -17,8 +17,8 @@ void Image_cuda_compatible::copy_to_GPU()
 {
     if(gpu_im == NULL)
         {
-        cudaMalloc( (void**)&gpu_im,size*sizeof(unsigned short));
-        cudaMemcpy(gpu_im,im,size * sizeof(unsigned short),cudaMemcpyHostToDevice);
+        cudaMalloc( (void**)&gpu_im,size*sizeof(float));
+        cudaMemcpy(gpu_im,im,size * sizeof(float),cudaMemcpyHostToDevice);
     }
 
 }
